+++ packages/force_inject/lib/src/service_descriptor.dart
@@ -8,11 +8,16 @@ class ServiceDescriptor {
   final Object? instance;
   final Object Function(ServiceProvider)? factory;
 
+  final Object? name;
+  final Object? tag;
+
   ServiceDescriptor({
     required this.type,
     required this.implementationType,
     required this.lifetime,
     this.instance,
     this.factory,
+    this.name,
+    this.tag,
   });
 }